made per_lap_snapshot module and moved entry class
